==> code/1. VC/1. Preprocessing(192021).ipynb <==
import warnings
warnings.filterwarnings('ignore')

import pandas as pd
import numpy as np

df = pd.read_csv('../data/preprocessing/data.csv')

df

#### 수정사항

df.rename(columns={'대기업': '매출구조비율(대기업)'}, inplace=True)
df.rename(columns={'12차벤더': '매출구조비율(12차벤더)'}, inplace=True)
df.rename(columns={'중소기업': '매출구조비율(중소기업)'}, inplace=True)
df.rename(columns={'B2C': '매출구조비율(B2C)'}, inplace=True)
df.rename(columns={'B2G': '매출구조비율(B2G)'}, inplace=True)
df.rename(columns={'글로벌': '매출구조비율(글로벌)'}, inplace=True)

pd.set_option('display.max_seq_items', None)
col_list = df.columns

df.rename(columns={'15_대표이사_이전근무지_법인': '대표이사_이전근무지'}, inplace=True)

# df.to_csv('data.csv', encoding='euc-kr')

df[df['인수합병 환경개선을 위한 필요사항(1순위)'] == 7.0]['인수합병 환경개선을 위한 필요사항(1순위)']

df.loc[df['인수합병 환경개선을 위한 필요사항(1순위)'] == 7.0, '인수합병 환경개선을 위한 필요사항(1순위)'] = np.nan

#### 학력별 인력구성 비율로 만들기

# 학력별 인력구성 컬럼 확인
columns_with_education = [col for col in df.columns if '학력별' in col]
columns_with_education

df['학력별 인력구성비(고졸이하)'] = (df['학력별 인력구성비(고졸이하)'] / df['학력별 인력구성_합계'])*100
df['학력별 인력구성비(전문대졸)'] = (df['학력별 인력구성비(전문대졸)'] / df['학력별 인력구성_합계'])*100
df['학력별 인력구성비(대졸)'] = (df['학력별 인력구성비(대졸)'] / df['학력별 인력구성_합계'])*100
df['학력별 인력구성비(석사)'] = (df['학력별 인력구성비(석사)'] / df['학력별 인력구성_합계'])*100
df['학력별 인력구성비(박사)'] = (df['학력별 인력구성비(박사)'] / df['학력별 인력구성_합계'])*100

df.rename(columns={'학력별 인력구성_고졸이하': '학력별 인력구성비(고졸이하)'}, inplace=True)
df.rename(columns={'학력별 인력구성_전문대졸': '학력별 인력구성비(전문대졸)'}, inplace=True)
df.rename(columns={'학력별 인력구성_대졸': '학력별 인력구성비(대졸)'}, inplace=True)
df.rename(columns={'학력별 인력구성_석사': '학력별 인력구성비(석사)'}, inplace=True)
df.rename(columns={'학력별 인력구성_박사': '학력별 인력구성비(박사)'}, inplace=True)

# df.to_csv('data.csv', encoding='euc-kr', index=False )

pd.set_option('display.max_seq_items', None)
df.columns

#### 부서별 인력구성비 만들기

df['인력구성(관리)'] = (df['인력구성(관리)'] / df['인력구성(합계)'])*100
df['인력구성(생산)'] = (df['인력구성(생산)'] / df['인력구성(합계)'])*100
df['인력구성(영업)'] = (df['인력구성(영업)'] / df['인력구성(합계)'])*100
df['인력구성(RnD)'] = (df['인력구성(RnD)'] / df['인력구성(합계)'])*100
df['인력구성(기타)'] = (df['인력구성(기타)'] / df['인력구성(합계)'])*100

df.rename(columns={'인력구성(관리)': '부서별_인력구성(관리)'}, inplace=True)
df.rename(columns={'인력구성(생산)': '부서별_인력구성(생산)'}, inplace=True)
df.rename(columns={'인력구성(영업)': '부서별_인력구성(영업)'}, inplace=True)
df.rename(columns={'인력구성(RnD)': '부서별_인력구성(RnD)'}, inplace=True)
df.rename(columns={'인력구성(기타)': '부서별_인력구성(기타)'}, inplace=True)

---

#### 기타 사항 수정하기

df.drop('크라우드 펀딩 경험여부', axis=1, inplace=True)
df.drop('회사주소_구(군)', axis=1, inplace=True)

df.rename(columns={'융자(정책자금)_경험여부': '융자(정책)_경험여부'}, inplace=True)

df.drop('연구개발(RnD)자금(출연금)_경험여부', axis=1, inplace=True)
df.drop('보증서지원_경험여부', axis=1, inplace=True)
df.drop('기타지원_경험여부', axis=1, inplace=True)

df.drop('인수합병의향', axis=1, inplace=True)
df.rename(columns={'인수합병응할의향' : '인수합병 고려 여부'}, inplace=True)

df.drop('4차산업혁명과관계여부', axis=1, inplace=True)

df.drop('현재 구인하기 어려운 직급(1순위)', axis=1, inplace=True)
df.drop('현재 구인하기 어려운 직급(2순위)', axis=1, inplace=True)

df.drop('경쟁사 대비 현황_개발역량', axis=1, inplace=True)
df.drop('경쟁사 대비 현황_제조역량', axis=1, inplace=True)
df.drop('경쟁사 대비 현황_마케팅역량', axis=1, inplace=True)
df.drop('경쟁전략초점_원가우위 수준', axis=1, inplace=True)
df.drop('경쟁전략초점_제품 고부가가치화 수준', axis=1, inplace=True)
df.drop('비즈니스 모델_명확한 정리정도', axis=1, inplace=True)
df.drop('비즈니스 모델_혁신정도', axis=1, inplace=True)

df.drop('벤처확인제도 혜택_창업 관련 규제 완화 효과', axis=1, inplace=True)
df.drop('벤처확인제도 혜택_입지 선정 시 유리', axis=1, inplace=True)
df.drop('현재 벤처인프라 수준_벤처투자 활성화', axis=1, inplace=True)

df.drop('향후사회적책임실천할계획여부', axis=1, inplace=True)

df['실천하고있는사회적책임'] = df['실천하고있는사회적책임'].replace([0.0, 1.0, 2.0, 3.0, 4.0, 5.0], 1)
df['실천하고있는사회적책임'] = df['실천하고있는사회적책임'].replace(6.0, 2)

df['실천하고있는사회적책임'].value_counts()

#### 산업 - 창업자 전공 매칭하기

df[['한국표준산업분류(KSIC)코드_10차']]

import re

df['산업'] = df['한국표준산업분류(KSIC)코드_10차'].apply(lambda x: re.sub('[^a-zA-Z]', '', x))

ind_list = df['산업'].unique()

ind_list

한국표준산업분류

* A : 농업, 임업 및 어업
* B : 광업
* C : 제조업
* D : 전기, 가스 증기 및 공기 조절 공급업
* E : 수도, 하수 및 폐기물 처리, 원료 재생업
* F : 건설업
* G : 도매 및 소매업
* H : 운수 및 창고업
* I : 숙박 및 음식점업
* J : 정보통신업
* K : 금융 및 보험업
* ~~L : 부동산업~~
* M : 전문, 과학 및 기술 서비스업
* N : 사업시설 관리, 사업 지원 및 임대 서비스업
* ~~O : 공공 행정, 국방 및 사회보장 행정~~
* P : 교육 서비스업
* Q : 보건업 및 사회복지 서비스업
* R : 예술, 스포츠 및 여가관련 서비스업
* S : 협회 및 단체, 수리 및 기타 개인 서비스업
* ~~T : 가구 내 고용활동 및 달리 분류되지 않은 자가 소비 생산~~
* ~~U : 국제 및 외국기관~~


창업자 전공

* 1 : 공학(엔지니어)
    * C, D, E, F, J
* 2 : 자연과학
    * A, B, D, E, M, 
* 3 : 경영/경제학
    * G, K, I, N, R, S, H
* 4 : 인문사회학
    * P, Q, 
* 5 : 기타

df.loc[(df['창업 당시 창업자 전공'] == 1) & (df['산업'].isin(['C', 'D', 'E', 'F', 'J'])), '전공과 업종 일치 여부'] = 1
df.loc[(df['창업 당시 창업자 전공'] == 2) & (df['산업'].isin(['A', 'B', 'D', 'E', 'M'])), '전공과 업종 일치 여부'] = 1
df.loc[(df['창업 당시 창업자 전공'] == 3) & (df['산업'].isin(['G', 'K', 'I', 'N', 'R', 'S', 'H'])), '전공과 업종 일치 여부'] = 1
df.loc[(df['창업 당시 창업자 전공'] == 4) & (df['산업'].isin(['P', 'Q'])), '전공과 업종 일치 여부'] = 1

df['전공과 업종 일치 여부'].fillna(2, inplace=True)

df['전공과 업종 일치 여부'].value_counts()

df.drop('창업 당시 창업자 전공', axis=1, inplace=True)
df.drop('산업', axis=1, inplace=True)

df

#### 컬럼명 비교

t_col_list = ['가중치',
 '벤처확인 유형',
 '회사주소_시(도)',
 '회사주소_구(군)',
 '한국표준산업분류(KSIC)코드_10차',
 '창업년도',
 '기업성장단계',
 '제조 및 비제조',
 '상장여부 및 향후계획',
 '대표이사_창업자여부',
 '대표이사_이전근무지',
 '창업자 성별',
 '창업 당시 창업자 연령',
 '창업 당시 창업자 최종학력',
 '창업 직전 근무지',
 '창업방식',
 '단독창업 여부',
 '창업자 역할',
 '창업 당시 목표시장',
 '창업 당시 창업자 실무경험 년수(년)',
 '창업자 과거창업 경험 여부',
 '학력별 인력구성비(고졸이하)',
 '학력별 인력구성비(전문대졸)',
 '학력별 인력구성비(대졸)',
 '학력별 인력구성비(석사)',
 '학력별 인력구성비(박사)',
 '현재 구인하기 어려운 직무(1순위)',
 '현재 구인하기 어려운 직무(2순위)',
 '지분구조(퍼센트)_창업자',
 '지분구조(퍼센트)_임직원',
 '지분구조(퍼센트)_가족',
 '지분구조(퍼센트)_벤처캐피털 및 기관투자자',
 '지분구조(퍼센트)_엔젤 및 액셀러레이터',
 '지분구조(퍼센트)_개인투자자',
 '융자(정책)_경험여부',
 '주력제품 국내 시장점유율(퍼센트)',
 '주력제품 해외 시장점유율(퍼센트)',
 '해외수출 및 진출과정 애로사항(1순위)',
 '해외수출 및 진출과정 애로사항(2순위)',
 '엔젤투자자 혹은 액셀러레이터 투자 경험여부',
 '벤처캐피털(투자조합포함) 투자 경험여부',
 '인수합병 고려 여부',
 '인수합병 환경개선을 위한 필요사항(1순위)',
 '인수합병 환경개선을 위한 필요사항(2순위)',
 '기업간 거래시 애로사항(1순위)',
 '벤처확인제도 혜택_기술개발 촉진효과',
 '벤처확인제도 혜택_기업 이미지 제고 등 광고효과',
 '벤처확인제도 혜택_세제혜택 효과',
 '벤처확인제도 혜택_우수인력 확보에 유리',
 '벤처확인제도 혜택_자금조달 및 투자 유인효과',
 '벤처확인제도 혜택_정부지원제도 참여에 유리',
 '현재 벤처인프라 수준_인수합병(MnA) 활성화를 위한 제도개선',
 '현재 벤처인프라 수준_벤처관련 규제완화',
 '현재 벤처인프라 수준_창업 및 기업가적 능력개발을 위한 교육',
 '현재 벤처인프라 수준_기업가정신을 고취하는 사회분위기',
 '경영상 애로사항_신기술 개발',
 '경영상 애로사항_개발된 기술의 사업화',
 '경영상 애로사항_국내 판로개척',
 '경영상 애로사항_해외시장 개척',
 '경영상 애로사항_필요인력의 확보 및 유지관리',
 '경영상 애로사항_자금조달_운용 등 자금관리',
 '경영상 애로사항_기술유출 및 디자인_상표 도용',
 '경영상 애로사항_업계내의 과당경쟁 및 덤핑',
 '경영상 애로사항_법률_회계_세무 관련 지식부족',
 '경영상 애로사항_조직관리',
 '경영상 애로사항_환율변동',
 '경영상 애로사항_특허분쟁',
 '경영상 애로사항_각종규제',
 '경영상 애로사항_산-학-연 간 협력활동',
 '경영상 애로사항_불공정한 시장질서',
 '매출규모',
 '정규직합계',
 '비정규직합계',
 '총종사자합계',
 '인력구성(관리)',
 '인력구성(생산)',
 '인력구성(영업)',
 '인력구성(RnD)',
 '인력구성(기타)',
 '인수합병(MnA) 경험여부',
 '피인수합병 경험여부',
 '인력구성(합계)',
 '학력별 인력구성_합계',
 '매출구조비율(대기업)',
 '매출구조비율(12차벤더)',
 '매출구조비율(중소기업)',
 '매출구조비율(B2C)',
 '매출구조비율(B2G)',
 '매출구조비율(글로벌)',
 '투자액(국내설비)',
 '투자액(R&D)',
 '투자액(해외)',
 '투자액(기타)',
 '투자액(합계)',
 '스톡옵션상황',
 '등록보유(특허권)',
 '등록보유(실용신안권)',
 '등록보유(디자인권)',
 '등록보유(상표권)',
 '등록보유(해외산업재산권)',
 '등록보유(합계)',
 '출원중(특허권)',
 '출원중(실용신안권)',
 '출원중(디자인권)',
 '출원중(상표권)',
 '출원중(해외산업재산권)',
 '출원중(합계)',
 '제품및서비스구조',
 '세계유일기술보유여부',
 '기술력수준비교(세계)',
 '기술력수준비교(국내)',
 '연구개발관련전담부서보유여부',
 '장기적성장전망성',
 '자체브랜드보유여부',
 '영업및마케팅방식',
 '글로벌사업방식',
 '실천하고있는사회적책임',
 '부채비율',
 '자기자본비율',
 '순이익률',
 '총자산회전율',
 '자기자본회전율',
 'ROE',
 '영업이익률',
 '이자보상배율',
 '종사자생산성',
 '전공과 업종 일치 여부']

result = [x for x in col_list if x not in t_col_list]

result

#### 스톡옵션상황 변경
1. 현재 활용 -> 1
2. 과거 활용, 현재 미활용 -> 2
3. 과거, 현재 미활용, 미래 활용 -> 2
4. 과거, 현재 미활용, 미래 미활용 -> 2

df['스톡옵션상황'].value_counts()

df['스톡옵션상황'] = df['스톡옵션상황'].replace([2, 3, 4], 2)
df['스톡옵션상황'] = df['스톡옵션상황'].replace(1, 1)

df['스톡옵션상황'].value_counts()

# df.to_csv('data192021.csv', encoding='euc-kr' , index=False)

# df.to_csv('data192021.csv', index=False)
