==> code/1. VC/2. Preprocessing(total).ipynb <==
# 6/19

import warnings
warnings.filterwarnings('ignore')

import pandas as pd
import numpy as np

df = pd.read_csv('../data/pre_/data.csv')

df['정규직합계'] = (df['정규직합계'] / df['총종사자합계']) * 100
df['비정규직합계'] = (df['비정규직합계'] / df['총종사자합계']) * 100

df.rename(columns={'정규직합계' : '정규직비율'}, inplace=True)
df.rename(columns={'비정규직합계' : '비정규직비율'}, inplace=True)

pd.reset_option('display.max_rows', None)
df[['정규직비율', '비정규직비율']]

df = df.round(3)

# df.to_csv('totaldata.csv', index=False, encoding='euc-kr')

---

# 6/20

df.rename(columns={'기업간 거래시 애로사항(1순위)' : '기업간 거래시 애로사항'}, inplace=True)

df['상장여부 및 향후계획'] = df['상장여부 및 향후계획'].astype(int)

df['대표이사_창업자여부'] = df['대표이사_창업자여부'].astype(int)

df['대표이사_이전근무지'] = df['대표이사_이전근무지'].astype(int, errors='ignore')

df.iloc[:, 10:19] = df.iloc[:, 10:19].astype(int, errors='ignore')

df.iloc[:, 24:26] = df.iloc[:, 24:26].astype(int, errors='ignore')

df['융자(정책)_경험여부'] = df['융자(정책)_경험여부'].astype(int, errors='ignore')

df.iloc[:,35:69] = df.iloc[:,35:69].astype(int, errors='ignore')

df['벤처확인 유형']

df[(df['벤처캐피털(투자조합포함) 투자 경험여부'] == 1) & (df['상장여부 및 향후계획'] == 1)]

df['벤처캐피털(투자조합포함) 투자 경험여부'].isna().sum()

---
데이터 수정

df['상장여부 및 향후계획'] = df['상장여부 및 향후계획'].replace([2,3,4,5], 2)
df['상장여부 및 향후계획'] = df['상장여부 및 향후계획'].replace(1, 1)

df['상장여부 및 향후계획'].value_counts()

df['창업 당시 창업자 최종학력'] = df['창업 당시 창업자 최종학력'].replace([5, 6], 5)

df['창업 당시 창업자 최종학력'].value_counts()

df.drop('현재 구인하기 어려운 직무(1순위)', axis=1, inplace=True)
df.drop('현재 구인하기 어려운 직무(2순위)', axis=1, inplace=True)

df['벤처캐피털(투자조합포함) 투자 경험여부'] = df['벤처캐피털(투자조합포함) 투자 경험여부'].replace([2,3,4], 2)

df['벤처캐피털(투자조합포함) 투자 경험여부'].value_counts()

df.drop('인수합병 고려 여부', axis=1, inplace=True)

df.drop('기업간 거래시 애로사항', axis=1, inplace=True)

df.drop('연구개발관련전담부서보유여부', axis=1, inplace=True)

df['영업및마케팅방식'] = df['영업및마케팅방식'].replace([4,5], 4)

df['인력구성(관리)'] = pd.to_numeric(df['인력구성(관리)'], errors='coerce')
df['인력구성(합계)'] = pd.to_numeric(df['인력구성(합계)'], errors='coerce')

df['인력구성(관리)'] = (df['인력구성(관리)'] / df['인력구성(합계)'])*100
df['인력구성(생산)'] = (df['인력구성(생산)'] / df['인력구성(합계)'])*100
df['인력구성(영업)'] = (df['인력구성(영업)'] / df['인력구성(합계)'])*100
df['인력구성(RnD)'] = (df['인력구성(RnD)'] / df['인력구성(합계)'])*100
df['인력구성(기타)'] = (df['인력구성(기타)'] / df['인력구성(합계)'])*100

df.rename(columns={'인력구성(관리)': '부서별_인력구성(관리)'}, inplace=True)
df.rename(columns={'인력구성(생산)': '부서별_인력구성(생산)'}, inplace=True)
df.rename(columns={'인력구성(영업)': '부서별_인력구성(영업)'}, inplace=True)
df.rename(columns={'인력구성(RnD)': '부서별_인력구성(RnD)'}, inplace=True)
df.rename(columns={'인력구성(기타)': '부서별_인력구성(기타)'}, inplace=True)

df
