==> code/1. VC/4-1. Oversampling(1,1).ipynb <==
import pandas as pd

df = pd.read_csv('../data/완성data/unsampled_data.csv')

print(df['벤처캐피털(투자조합포함) 투자 경험여부'].value_counts().sum())
print(df['벤처캐피털(투자조합포함) 투자 경험여부'].value_counts())

## OverSampling

# 컬럼 위치 확인
column_name = '가중치'
column_index = df.columns.tolist().index(column_name)
column_index

# 컬럼 위치 확인
column_name = '경영상 애로사항(개수)'
column_index = df.columns.tolist().index(column_name)
column_index

# 분석할 데이터 설정
data = df.iloc[:,5:66]

# 원본 데이터 (12384 개)
from imblearn.over_sampling import SMOTE

X = data.iloc[:, :-1]
y = data['벤처캐피털(투자조합포함) 투자 경험여부']
print(X.shape)

# SMOTE

smote = SMOTE(sampling_strategy=0.5, random_state=0)
X_resampled, y_resampled = smote.fit_resample(X, y)
print(X_resampled.shape)

# 오버샘플링된 데이터를 데이터프레임으로 변환
df_resampled = pd.DataFrame(X_resampled, columns=X.columns)
df_resampled['벤처캐피털(투자조합포함) 투자 경험여부'] = y_resampled

# 데이터프레임 저장
# df_resampled.to_csv('oversampled_data.csv', index=False)
