==> code/1. VC/5. Encoding.ipynb <==
import warnings
warnings.filterwarnings('ignore')

import pandas as pd
import numpy as np

df = pd.read_csv('../data/pre_/preprocessing_data.csv')

df

pd.set_option('display.max_seq_items', None)
df.columns

### Dtype 변환

df['인수합병(MnA) 경험여부'] = df['인수합병(MnA) 경험여부'].astype('Int64')
df['대표이사_창업자여부'] = df['대표이사_창업자여부'].astype('Int64')
df['단독창업 여부'] = df['단독창업 여부'].astype('Int64')
df['창업 당시 목표시장'] = df['창업 당시 목표시장'].astype('Int64')
df['융자(정책)_경험여부'] = df['융자(정책)_경험여부'].astype('Int64')
df['장기적성장전망성'] = df['장기적성장전망성'].astype('Int64')
df['융자(정책)_경험여부'] = df['융자(정책)_경험여부'].astype('Int64')
df['전공과 업종 일치 여부'] = df['전공과 업종 일치 여부'].astype('Int64')
df['연구개발관련전담부서보유여부'] = df['연구개발관련전담부서보유여부'].astype('Int64')
df['해외수출 및 진출과정 애로사항'] = df['해외수출 및 진출과정 애로사항'].astype('Int64')
df['제품및서비스구조'] = df['제품및서비스구조'].astype('Int64')
df['기술력수준비교(국내)'] = df['기술력수준비교(국내)'].astype('Int64')
df['기술력수준비교(세계)'] = df['기술력수준비교(세계)'].astype('Int64')

### 1, 2 -> 0, 1 로 변환

df['상장여부'] = df['상장여부'].replace([2], 0)
df['대표이사_창업자여부'] = df['대표이사_창업자여부'].replace([2], 0)
df['단독창업 여부'] = df['단독창업 여부'].replace([2], 0)
df['창업자 과거창업 경험 여부'] = df['창업자 과거창업 경험 여부'].replace([2], 0)
df['융자(정책)_경험여부'] = df['융자(정책)_경험여부'].replace([2], 0)
df['자체브랜드보유여부'] = df['자체브랜드보유여부'].replace([2], 0)
df['전공과 업종 일치 여부'] = df['전공과 업종 일치 여부'].replace([2], 0)
df['엔젤투자자 혹은 액셀러레이터 투자 경험여부'] = df['엔젤투자자 혹은 액셀러레이터 투자 경험여부'].replace([2], 0)

df['인수합병(MnA) 경험여부'] = df['인수합병(MnA) 경험여부'].replace([2], 0)
df['인수합병(MnA) 경험여부'] = df['인수합병(MnA) 경험여부'].replace([3], 2)

### 컬럼 위치 변경

new_order = ['회사주소_시(도)', '한국표준산업분류(KSIC)코드_10차', '벤처확인 유형', '상장여부', '인수합병(MnA) 경험여부',
        '가중치',  '기업성장단계',
       '제조 및 비제조', '대표이사_창업자여부', '창업 당시 창업자 연령', '창업 당시 창업자 최종학력',
       '창업 직전 근무지', '창업방식', '단독창업 여부', '창업 당시 목표시장', 
       '창업자 과거창업 경험 여부', '융자(정책)_경험여부', '장기적성장전망성', '자체브랜드보유여부', '영업및마케팅방식',
       '전공과 업종 일치 여부', '연구개발관련전담부서보유여부', '엔젤투자자 혹은 액셀러레이터 투자 경험여부',
       '해외수출 및 진출과정 애로사항', '제품및서비스구조', '기술력수준비교(세계)',
       '기술력수준비교(국내)', '주력제품 국내 시장점유율(퍼센트)', '주력제품 해외 시장점유율(퍼센트)', '정규직비율',
       '학력별 인력구성비(고졸이하)', '학력별 인력구성비(전문대졸)', '학력별 인력구성비(대졸)', '학력별 인력구성비(석사)',
       '학력별 인력구성비(박사)', '부서별_인력구성(관리)', '부서별_인력구성(생산)', '부서별_인력구성(영업)',
       '부서별_인력구성(RnD)', '부서별_인력구성(기타)', '매출구조비율(대기업)', '매출구조비율(12차벤더)',
       '매출구조비율(중소기업)', '매출구조비율(B2C)', '매출구조비율(B2G)', '매출구조비율(글로벌)',
       '투자액비(국내설비)', '투자액비(RnD)', '투자액비(해외)', '투자액비(기타)', '부채비율', '자기자본비율',
       '순이익률', '총자산회전율', '자기자본회전율', 'ROE', '영업이익률', '이자보상배율', '종사자생산성',
       '창업 당시 창업자 실무경험 년수(년)', '벤처확인제도 혜택(평균)', '벤처인프라 수준(평균)', '등록보유(합계)', '출원중(합계)',
        '경영상 애로사항(개수)', 
        '벤처캐피털(투자조합포함) 투자 경험여부']

df = df.reindex(columns=new_order)

### One-Hot Encoding

* 기업성장단계 -> 5
* 창업 당시 창업자 연령 -> 5
* 제조 및 비제조 -> 4
* 창업 당시 창업자 최종학력 -> 5
* 창업 직전 근무지 -> 10
* 창업방식 -> 3
* 창업 당시 목표시장 -> 3
* 장기적성장전망성 -> 5
* 영업및마케팅방식 -> 4
* 연구개발관련전담부서보유여부 -> 4
* 해외수출 및 진출과정 애로사항 -> 8
* 제품및서비스구조 -> 3
* 기술력수준비교(세계) & (국내)) -> 5

# 원-핫 인코딩 적용
df_encoded = pd.get_dummies(df, columns=['기업성장단계', '창업 당시 창업자 연령', '제조 및 비제조', '창업 당시 창업자 최종학력',
                                           '창업 직전 근무지', '창업방식', '창업 당시 목표시장', '장기적성장전망성', '영업및마케팅방식',
                                           '연구개발관련전담부서보유여부', '해외수출 및 진출과정 애로사항', '제품및서비스구조',
                                           '기술력수준비교(세계)','기술력수준비교(국내)'])

df_encoded.to_csv('encoded_data.csv', index=False, encoding='utf-8')
