==> code/1. VC/6. Scaling.ipynb <==
import pandas as pd
from sklearn.preprocessing import StandardScaler, MinMaxScaler, RobustScaler, MaxAbsScaler

### Scaler

* ~~StandardScaler~~
    - 특성들의 평균을 0, 분산을 1로 스케일링하는 것 (특성들을 정규분포로)
    - 이상치에 민감
    - 회귀보다 분류에 유용
* MinMaxScaler
    - 가장 작은 값은 0, 가장 큰 값은 1로 변환되므로, 모든 특성들은 0 ~ 1 범위를 갖는다
    - 이상치에 민감
    - 분류보다 회귀에 유용
* ~~RobustScaler~~
    - 평균과 분산 대신에 중간값과 사분위값 사용
    - 이상치 영향을 최소화
* ~~MaxAbsScaler~~
    - 각 특성의 절대값이 0과 1사이가 되도록 스케일링
    - 모든 값이 -1 ~ 1 사이로 표현되며, 데이터가 양수일 경우 MinMaxScaler 와 동일
    - 이상치에 민감

df = pd.read_csv('../data/완성data/최종사용data/oversampled_data.csv')

pd.set_option('display.max_seq_items', None)
df.columns

#### 비율형 변수만 분리

# 컬럼 위치 확인
column_name = '주력제품 국내 시장점유율(퍼센트)'
column_index = df.columns.tolist().index(column_name)
column_index

column_name = '경영상 애로사항(개수)'
column_index = df.columns.tolist().index(column_name)
column_index

df.iloc[:,22:60].columns

tran_df = df.iloc[:,22:60]

#### 분리한 데이터 Scaling

# Minmaxscaler -> 인코딩된 범주형 변수 값인 0, 1 값과 유사한 범위 값으로 바꿔주기 위해서
mm = MinMaxScaler()
mm_tran_df = mm.fit_transform(tran_df)
mm_tran_df = pd.DataFrame(mm_tran_df, columns=tran_df.columns)

#### 원본 데이터를 Scaling 된 데이터로 치환

df.iloc[:,22:60] = mm_tran_df

df.iloc[:,22:60]

# df.to_csv('mmscaled_data.csv', index=False, encoding='utf-8')
