==> code/1. VC/9. Cross_validation.ipynb <==
from sklearn.model_selection import StratifiedKFold
from sklearn.ensemble import RandomForestClassifier
from sklearn.metrics import accuracy_score, f1_score
import pandas as pd
import numpy as np

data = pd.read_csv('../data/완성data/unsampled_data.csv')

# 분석할 데이터 설정
df = data.iloc[:,5:66]

df

df = df[['벤처확인 유형',
 '기업성장단계',
 '제조 및 비제조',
 '대표이사_창업자여부',
 '창업 당시 창업자 연령',
 '창업 당시 창업자 최종학력',
 '창업 직전 근무지',
 '창업 당시 목표시장',
 '장기적성장전망성',
 '자체브랜드보유여부',
 '엔젤투자자 혹은 액셀러레이터 투자 경험여부',
 '해외수출 및 진출과정 애로사항',
 '제품및서비스구조',
 '기술력수준비교(세계)',
 '기술력수준비교(국내)',
 '벤처캐피털(투자조합포함) 투자 경험여부']]

# Stratified K-Fold

skf = StratifiedKFold(n_splits=5, shuffle=True, random_state=0)
n_iter = 0

for train_index, test_index in skf.split(df, df['벤처캐피털(투자조합포함) 투자 경험여부']):
    train_data = df.iloc[train_index]
    test_data = df.iloc[test_index]

    print('## 교차검증 : {0}'.format(n_iter))
    print('학습 레이블 데이터 분포 : \n', train_data.value_counts())
    print('검증 레이블 데이터 분포 : \n', test_data.value_counts())

# 훈련 데이터에서 독립변수와 타겟변수를 나누기
X_train = train_data.drop('벤처캐피털(투자조합포함) 투자 경험여부', axis=1)
y_train = train_data['벤처캐피털(투자조합포함) 투자 경험여부']

# 테스트 데이터에서 독립변수와 타겟변수를 나누기
X_test = test_data.drop('벤처캐피털(투자조합포함) 투자 경험여부', axis=1)
y_test = test_data['벤처캐피털(투자조합포함) 투자 경험여부']

# RandomForest 모델 생성 및 훈련
from sklearn.linear_model import LogisticRegression
model = LogisticRegression(random_state=0)
model.fit(X_train, y_train)

# 테스트 데이터로 예측 수행
y_pred = model.predict(X_test)

# 평가지표 계산
accuracy = accuracy_score(y_test, y_pred)
F1_score = f1_score(y_test, y_pred)
print('Accuracy : ', accuracy)
print('F1 score : ', F1_score)

# RandomForest 모델 생성 및 훈련
model = RandomForestClassifier(random_state=0)
model.fit(X_train, y_train)

# 테스트 데이터로 예측 수행
y_pred = model.predict(X_test)

# 평가지표 계산
accuracy = accuracy_score(y_test, y_pred)
F1_score = f1_score(y_test, y_pred)
print('Accuracy : ', accuracy)
print('F1 score : ', F1_score)

from sklearn.metrics import confusion_matrix, recall_score, precision_score, roc_auc_score

a = confusion_matrix(y_test, y_pred)
a

# Cross_val_score

from sklearn.model_selection import cross_val_score, cross_validate


# 독립변수와 종속변수 설정
X = df.iloc[:, :-1]
y = df['벤처캐피털(투자조합포함) 투자 경험여부']

# 모델 생성
model = RandomForestClassifier()

# 교차검증 수행
cv = StratifiedKFold(n_splits=5, shuffle=True, random_state=0)
scores = cross_val_score(model, X, y, scoring='f1', cv = cv)
accuracy_scores = cross_val_score(model, X, y, scoring='accuracy', cv=cv)

# 교차검증 결과 출력
print('Cross-Validation F1 scores : ', scores)
print('Mean F1 score : ', scores.mean())
print('Cross-Validation Accuracy scores:', accuracy_scores)
print('Mean Accuracy:', accuracy_scores.mean())
